+++ e2e/tests/nexus--edgecases-import_from_same_pkg--google-protobuf/schema.ts
@@ -1,5 +1,5 @@
 import { makeTestSchema } from "@proto-graphql/e2e-helper";
-import { ParentMessage } from "@testapis/node-native/lib/testapis/edgecases/import_from_same_pkg/parent_pb";
+import { ParentMessage } from "@proto-graphql/e2e-testapis-google-protobuf/lib/testapis/edgecases/import_from_same_pkg/parent_pb";
 import { queryField } from "nexus";
 
 import * as childTypes from "./__generated__/schema/testapis/edgecases/import_from_same_pkg/child_pb_nexus";