+++ e2e/tests/nexus--empty_types--protobufjs/schema.ts
@@ -1,5 +1,5 @@
 import { makeTestSchema } from "@proto-graphql/e2e-helper";
-import * as pbjs from "@testapis/node/lib/testapis/empty_types";
+import * as pbjs from "@proto-graphql/e2e-testapis-protobufjs/lib/testapis/empty_types";
 import { queryField } from "nexus";
 
 import * as types1 from "./__generated__/schema/testapis/empty_types/empty_pb_nexus";